Moved printf to kernel
Fixes:
  - None

Featurs:
  - printf moved to kernel directory. It makes more sense
    there since it is OSAL independent.
