+++ src/kernel/printf.c
@@ -33,7 +33,7 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-#include <osal/printf.h>
+#include <kernel/printf.h>
 
 // define this globally (e.g. gcc -DPRINTF_INCLUDE_CONFIG_H ...) to include the
 // printf_config.h header file
